+++ src/const.js
@@ -56,4 +56,14 @@ const DateFormat = {
   DATE_PICKER: 'd/m/y H:i',
 };
 
-export { SortType, EVENT_TYPES, defaultPoint, UpdateType, UserAction, timeType, DateFormat };
+const TimeLimit = {
+  LOWER_LIMIT: 350,
+  UPPER_LIMIT: 1000,
+};
+
+const ModeAdded = {
+  DEFAULT: 'DEFAULT',
+  ADDED: 'ADDED',
+};
+
+export { SortType, EVENT_TYPES, defaultPoint, UpdateType, UserAction, timeType, DateFormat,TimeLimit,ModeAdded };
